+++ RockPaperScissors/RockPaperScissors/ChoiceSelection.py
@@ -1,12 +1,13 @@
 from tkinter import *;
 import Simulation;
+import ImagesHelper
 from PIL import Image, ImageTk
 
 class ChoiceSelectionForm(Toplevel):
     def __init__(self,master=None):
         super().__init__(master=master)
         self.title("Choice Selection")
-        self.geometry("261x200")
+        self.geometry("350x200")
         choiceLabel = Label(self,text="Please select which you will use")
         choiceLabel.grid(row=0,column=1);
 
@@ -14,9 +15,9 @@ class ChoiceSelectionForm(Toplevel):
         global paperImage
         global scissorsImage
 
-        rockImage = GetImage("Rock")
-        paperImage = GetImage("Paper")
-        scissorsImage = GetImage("Scissors")
+        rockImage = ImagesHelper.GetImage("Rock")
+        paperImage = ImagesHelper.GetImage("Paper")
+        scissorsImage = ImagesHelper.GetImage("Scissors")
         
 
         rockButton = Button(self,text="Rock",image=rockImage);
@@ -33,20 +34,3 @@ class ChoiceSelectionForm(Toplevel):
         scissorsButton.bind("<Button>",
                         lambda e: Simulation.ScissorsSimulationForm(self))
         scissorsButton.grid(row=1,column=2)
-
-def GetImage(choice):
-    if choice == "Rock":
-        imageConverter = Image.open("Images/Rock.png")
-        imageConverter = imageConverter.resize((50,50), Image.ANTIALIAS)
-
-        return ImageTk.PhotoImage(imageConverter)
-    elif choice == "Paper":
-        imageConverter = Image.open("Images/Paper.png")
-        imageConverter = imageConverter.resize((50,50), Image.ANTIALIAS)
-
-        return ImageTk.PhotoImage(imageConverter)
-    elif choice == "Scissors":
-        imageConverter = Image.open("Images/Scissors.png")
-        imageConverter = imageConverter.resize((50,50), Image.ANTIALIAS)
-
-        return ImageTk.PhotoImage(imageConverter)
